+++ common/src/main/java/com/gregtechceu/gtceu/api/gui/GuiTextures.java
@@ -1,7 +1,9 @@
 package com.gregtechceu.gtceu.api.gui;
 
+import com.gregtechceu.gtceu.api.GTValues;
 import com.lowdragmc.lowdraglib.gui.texture.ResourceBorderTexture;
 import com.lowdragmc.lowdraglib.gui.texture.ResourceTexture;
+import lombok.val;
 
 /**
  * @author KilaBash
@@ -59,6 +61,8 @@ public class GuiTextures {
     public static final ResourceTexture BUTTON_CHUNK_MODE = new ResourceTexture("gtceu:textures/gui/widget/button_chunk_mode.png");
     public static final ResourceTexture BUTTON_CLEAR_GRID = new ResourceTexture("gtceu:textures/gui/widget/button_clear_grid.png");
     public static final ResourceTexture BUTTON_FILTER_DAMAGE = new ResourceTexture("gtceu:textures/gui/widget/button_filter_damage.png");
+    public static final ResourceTexture BUTTON_DISTINCT_BUSES = new ResourceTexture("gtceu:textures/gui/widget/button_distinct_buses.png");
+    public static final ResourceTexture BUTTON_POWER = new ResourceTexture("gtceu:textures/gui/widget/button_power.png");
     public static final ResourceTexture BUTTON_FILTER_NBT = new ResourceTexture("gtceu:textures/gui/widget/button_filter_nbt.png");
     public static final ResourceTexture BUTTON_FLUID_OUTPUT = new ResourceTexture("gtceu:textures/gui/widget/button_fluid_output_overlay.png");
     public static final ResourceTexture BUTTON_ITEM_OUTPUT = new ResourceTexture("gtceu:textures/gui/widget/button_item_output_overlay.png");
@@ -274,4 +278,11 @@ public class GuiTextures {
     public static final ResourceTexture LIGHT_ON = new ResourceTexture("gtceu:textures/gui/widget/light_on.png");
     public static final ResourceTexture LIGHT_OFF = new ResourceTexture("gtceu:textures/gui/widget/light_off.png");
     public static final ResourceTexture UP = new ResourceTexture("gtceu:textures/gui/base/up.png");
+    public static final ResourceTexture[] TIER = new ResourceTexture[9];
+    static {
+        val offset = 1f / TIER.length;
+        for (int i = 0; i < TIER.length; i++) {
+            TIER[i] = new ResourceTexture("gtceu:textures/gui/overlay/tier.png").getSubTexture(0, i * offset, 1, offset);
+        }
+    }
 }