+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # ChangeLog
 
-* fix kjs compatibility
-* fix gas collector doesnt work
+* fix dedicated server crash
+* fix missing textures
+* fix custom kjs builders
+* fix shaped recipe for mold,dust, etc
+* add mining hammer
+* add robot arm
+* add processing array
+* add distinct logic
+* add machine on/off button
+* improve fluid cell and prospector rendering